requirements.txt
tensorflow==1.15.2
opencv-python
kivy
pillow
firebase
base64
--upgrade pip wheel setuptools 
docutils pygments pypiwin32 kivy.deps.sdl2 kivy.deps.glew 
kivy.deps.gstreamer 
kivy.deps.angle 
pygame
mysql
mysql-connector-python-rf